requirements.txt
httpx>=0.25.2
pydantic>=2.5.3
loguru>=0.7.2
argon2-cffi>=21.3.0
msgpack>=1.1.0
fastapi>=0.109.2
uvicorn>=0.27.1
python-dotenv>=1.0.0

